CM-145: added Title on Gatsby
